core: extract ldelf related code from user_ta.c
Moves ldelf functionality from user_ta.c to a separate file.
This is the first step for decoupling ldelf from user TAs.
