+++ core/arch/arm/kernel/sub.mk
@@ -1,5 +1,6 @@
 ifeq ($(CFG_WITH_USER_TA),y)
 srcs-y += user_ta.c
+srcs-y += ldelf_loader.c
 srcs-$(CFG_REE_FS_TA) += ree_fs_ta.c
 srcs-$(CFG_EARLY_TA) += early_ta.c
 srcs-$(CFG_SECSTOR_TA) += secstor_ta.c
